fastapi_app/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
joblib
numpy
pandas
azure-eventhub
scikit-learn
xgboost
azure-storage-blob
jinja2